Create matchings_plot.R
This script does EB and CEM using the template of the authors, so we can easily reproduce the same plot as them. The results are pretty striking, since effects disappear or even change sign, we should look at that more closely (due to the logic of matching, it doesn't seem something unfeasible).